feat: enhance file reading security by validating workspace root; add tests for blog page logic
